Added build.gradle to /core project.
The goal is to make it easier for Android projects (which are now
gravitating towards gradle instead of ant) to depend on the 'core'
of zxing, as a source dependency, rather than a .jar.

For this reason, only the core directory had a build script added.
It would be nice to also include build scripts for android-integration,
but that will be something that can be done in the future.